Fix tests by combining parsers into a single test suite.
This makes the tests more DRY by moving all the test parsers into a
single test suite and using a test generator to parametrize the tests.
